refactoring for the ScreenshotViewer + fix it to not be raycast target
