+++ unity-renderer/Assets/Features/CameraReel/ScreenshotViewer/CameraReelSectionView.cs
@@ -1,16 +1,9 @@
-using Cysharp.Threading.Tasks;
+using CameraReel.ScreenshotViewer;
 using DCL;
 using DCLServices.CameraReelService;
-using DCLServices.WearablesCatalogService;
-using System;
 using System.Collections;
-using System.Collections.Generic;
-using System.Threading;
-using TMPro;
-using UI.InWorldCamera.Scripts;
 using UnityEngine;
 using UnityEngine.Networking;
-using UnityEngine.Serialization;
 using UnityEngine.UI;
 using Environment = DCL.Environment;
 
@@ -18,91 +11,24 @@ public class CameraReelSectionView : MonoBehaviour
 {
     private const int LIMIT = 20;
 
-    private readonly List<GameObject> profiles = new ();
-    [Header("Gallery view")]
+    [Header("GALLERY VIEW")]
     [SerializeField] private Button showMore;
     [SerializeField] internal Image prefab;
     [SerializeField] internal GridContainerComponentView gridContrainer;
 
-    [SerializeField] internal ScrollRect scrollView;
-    [SerializeField] private Canvas canvas;
+    [Space(10)]
+    [SerializeField] private ScreenshotViewerHUDView screenshotViewer;
 
-    [Header("Screenshot View")]
-    [SerializeField] private GameObject screenShotView;
-    [SerializeField] private Image screenImage;
-    [SerializeField] private Button closeScreenshotView;
-    [SerializeField] private TMP_Text dataTime;
-    [SerializeField] private TMP_Text sceneInfo;
-
-    [FormerlySerializedAs("profileEntry")] [SerializeField] internal ScreenshotVisiblePersonView profileEntryTemplate;
-    [SerializeField] internal Transform profileGridContrainer;
-    [SerializeField] private NFTItemInfo nftItemInfo;
-
-    [SerializeField] internal Button downloadButton;
-    [SerializeField] internal Button deleteButton;
-    [SerializeField] internal Button linkButton;
-    [SerializeField] internal Button twitterButton;
-
-    private Canvas gridCanvas;
     private int offset;
-
-    private CameraReelResponse currentScreenshot;
-
-    public void Awake()
-    {
-        gridCanvas = gridContrainer.GetComponent<Canvas>();
-        profileEntryTemplate.gameObject.SetActive(false);
-    }
 
     private void OnEnable()
     {
         showMore.onClick.AddListener(LoadImages);
-        closeScreenshotView.onClick.AddListener(() => screenShotView.SetActive(false));
-        downloadButton.onClick.AddListener(Download);
-
-        deleteButton.onClick.AddListener(DeleteImage);
-        linkButton.onClick.AddListener(CopyLink);
-        twitterButton.onClick.AddListener(CopyTwitterLink);
     }
 
     private void OnDisable()
     {
         showMore.onClick.RemoveAllListeners();
-        downloadButton.onClick.RemoveAllListeners();
-        closeScreenshotView.onClick.RemoveAllListeners();
-
-        deleteButton.onClick.RemoveAllListeners();
-        linkButton.onClick.RemoveAllListeners();
-        twitterButton.onClick.RemoveAllListeners();
-    }
-
-    private void CopyTwitterLink()
-    {
-        var description = "Check out what I'm doing in Decentraland right now and join me!";
-        var url = $"https://dcl.gg/reels?image={currentScreenshot.id}";
-        var twitterUrl = $"https://twitter.com/intent/tweet?text={description}&hashtags=DCLCamera&url={url}";
-
-        GUIUtility.systemCopyBuffer = twitterUrl;
-        Application.OpenURL(twitterUrl);
-    }
-
-    private void CopyLink()
-    {
-        var url = $"https://dcl.gg/reels?image={currentScreenshot.id}";
-
-        GUIUtility.systemCopyBuffer = url;
-        Application.OpenURL(url);
-    }
-
-    private void Download()
-    {
-        Application.OpenURL(currentScreenshot.url);
-    }
-
-    private async void DeleteImage()
-    {
-        ICameraReelNetworkService cameraReelNetworkService = Environment.i.serviceLocator.Get<ICameraReelNetworkService>();
-        await cameraReelNetworkService.DeleteScreenshot(currentScreenshot.id);
     }
 
     private async void LoadImages()
@@ -133,127 +59,13 @@ public class CameraReelSectionView : MonoBehaviour
                 image.gameObject.SetActive(true);
 
                 Button button = image.GetComponent<Button>();
-                button.onClick.AddListener(() => StartCoroutine(ShowScreenshotWithMetadata(reel)));
+                button.onClick.AddListener(() => ShowScreenshotWithMetadata(reel));
             }
         }
     }
 
-    private IEnumerator ShowScreenshotWithMetadata(CameraReelResponse reel)
+    private void ShowScreenshotWithMetadata(CameraReelResponse reel)
     {
-        currentScreenshot = reel;
-        screenShotView.SetActive(true);
-
-        // Show Screenshot
-        {
-            UnityWebRequest request = UnityWebRequestTexture.GetTexture(reel.thumbnailUrl);
-            yield return request.SendWebRequest();
-
-            if (request.result != UnityWebRequest.Result.Success)
-                Debug.Log(request.error);
-            else
-            {
-                sceneInfo.text = $"{reel.metadata.scene.name}, {reel.metadata.scene.location.x}, {reel.metadata.scene.location.y}";
-
-                Texture2D texture = ((DownloadHandlerTexture)request.downloadHandler).texture;
-                screenImage.sprite = Sprite.Create(texture, new Rect(0, 0, texture.width, texture.height), new Vector2(0.5f, 0.5f));
-            }
-        }
-
-        // Show DateTime Metadata
-        if (long.TryParse(reel.metadata.dateTime, out long unixTimestamp))
-        {
-            var epoch = new DateTime(1970, 1, 1, 0, 0, 0, DateTimeKind.Utc);
-            dataTime.text = epoch.AddSeconds(unixTimestamp).ToLocalTime().ToString("MMMM dd, yyyy");
-        }
-
-        // Show Visible Persons
-        foreach (GameObject profileGameObject in profiles)
-            Destroy(profileGameObject);
-
-        profiles.Clear();
-
-
-        foreach (VisiblePerson visiblePerson in reel.metadata.visiblePeople)
-        {
-            ScreenshotVisiblePersonView profileEntry = Instantiate(this.profileEntryTemplate, profileGridContrainer);
-            ProfileCardComponentView profile = profileEntry.ProfileCard;
-            profile.SetProfileName(visiblePerson.userName);
-            profile.SetProfileAddress(visiblePerson.userAddress);
-
-            profiles.Add(profileEntry.gameObject);
-
-            UpdateProfileIcon(visiblePerson.userAddress, profile);
-
-            if(visiblePerson.wearables.Length > 0)
-            {
-                profileEntry.WearablesListContainer.gameObject.SetActive(true);
-                IWearablesCatalogService wearablesService = Environment.i.serviceLocator.Get<IWearablesCatalogService>();
-                FetchWearables(visiblePerson, wearablesService, profileEntry);
-            }
-
-            profileEntry.gameObject.SetActive(true);
-        }
-    }
-
-    private async void UpdateProfileIcon(string userId, ProfileCardComponentView person)
-    {
-        UserProfile profile = UserProfileController.userProfilesCatalog.Get(userId) ?? await UserProfileController.i.RequestFullUserProfileAsync(userId);
-
-        UnityWebRequest request = UnityWebRequestTexture.GetTexture(profile.face256SnapshotURL);
-        await request.SendWebRequest();
-
-        if (request.result != UnityWebRequest.Result.Success)
-            Debug.Log(request.error);
-        else
-        {
-            Texture2D texture = ((DownloadHandlerTexture)request.downloadHandler).texture;
-            person.SetProfilePicture(Sprite.Create(texture, new Rect(0, 0, texture.width, texture.height), new Vector2(0.5f, 0.5f)));
-        }
-    }
-
-    private async void FetchWearables(VisiblePerson person, IWearablesCatalogService wearablesService, ScreenshotVisiblePersonView profileEntry)
-    {
-        foreach (string wearable in person.wearables)
-        {
-            WearableItem wearableItem = await wearablesService.RequestWearableAsync(wearable, default(CancellationToken));
-
-            NFTIconComponentView wearableEntry = Instantiate(profileEntry.WearableTemplate, profileEntry.WearablesListContainer);
-            NFTIconComponentModel newModel = new NFTIconComponentModel
-                {
-                    name = wearableItem.GetName(),
-                    imageURI = wearableItem.ComposeThumbnailUrl(),
-                    rarity = wearableItem.rarity,
-                    nftInfo = wearableItem.GetNftInfo(),
-                    marketplaceURI = wearableItem.GetMarketplaceLink(),
-                    showMarketplaceButton = true,
-                    showType = false,
-                    type = wearableItem.data.category,
-                };
-
-            wearableEntry.Configure(newModel);
-            // // Show Screenshot
-            // {
-            //     UnityWebRequest request = UnityWebRequestTexture.GetTexture(wearableItem.ComposeThumbnailUrl());
-            //
-            //     await request.SendWebRequest();
-            //
-            //     if (request.result != UnityWebRequest.Result.Success)
-            //         Debug.Log(request.error);
-            //     else
-            //     {
-            //         Texture2D texture = ((DownloadHandlerTexture)request.downloadHandler).texture;
-            //         var sprite = Sprite.Create(texture, new Rect(0, 0, texture.width, texture.height), new Vector2(0.5f, 0.5f));
-            //
-            //         newModel.imageURI = wearableItem.ComposeThumbnailUrl();
-            //         newModel.isOwned = true;
-            //
-            //
-            //         wearableEntry.Configure(newModel);
-            //             ConfigureNFTItemInfo(nftItemInfo, wearableItem, true);
-            //     }
-            // }
-
-            wearableEntry.gameObject.SetActive(true);
-        }
+        screenshotViewer.Show(reel);
     }
 }
