+++ unity-renderer/Assets/Features/CameraReel/ScreenshotViewer/Scripts/ScreenshotViewerHUDController.cs
@@ -1,0 +1,7 @@
+﻿namespace CameraReel.ScreenshotViewer
+{
+    public class ScreenshotViewerHUDController
+    {
+
+    }
+}
