+++ unity-renderer/Assets/Features/CameraReel/ScreenshotViewer/Scripts/ScreenshotVisiblePersonView.cs
